Move notifications page view models to dedicated package
