+++ app/views/components/notifications_single_element.scala.html
@@ -14,7 +14,7 @@
  * limitations under the License.
  *@
 
-@import models.viewmodels.NotificationsPageSingleElement
+@import models.viewmodels.notificationspage.NotificationsPageSingleElement
 
 @(index: Int, element: NotificationsPageSingleElement)
