+++ app/views/notifications.scala.html
@@ -14,7 +14,7 @@
  * limitations under the License.
  *@
 
-@import models.viewmodels.NotificationsPageSingleElement
+@import models.viewmodels.notificationspage.NotificationsPageSingleElement
 
 @this(main_template: views.html.main_template)
